requirements.txt
black>=26.1.0
email-validator>=2.3.0
pre-commit>=4.5.1
fastapi>=0.128.8
httpx>=0.28.0
isort>=7.0.0
python-decouple>=3.8
pytest>=9.0.2
uvicorn[standard]>=0.40.0
